ADK_SAAS_Agent/backend/requirements.txt
fastapi>=0.115.0
uvicorn>=0.34.0
google-adk[all]>=0.1.0
firebase-admin>=6.4.0
google-cloud-firestore>=2.14.0
google-genai>=0.1.1
slowapi>=0.1.9
